refactor topGenres
